Some more comments and suggestions
Line 1215: it says: 1) set the `type` key to the window plugin name, this is typically done by setting the window plugin as one of the `dependencies` 
How can we link plugin names and their URL in the dependencies? I would expect to have the dependencies as follows, but then how should I refer to it in `type`?
```
"dependencies": ["https://github.com/imjoy-team/imjoy-plugins/blob/master/repository/mylovelyplugin.imjoy.html"]
``` 

Line 1217
Sorry, I don't understand why you need to run both cells and how they connect to each other. It says that we're calling a Python function from JavaScript but I don't really see it.

Line 1282: It sais that we can write  `data={"process": self.process, "_rintf": True}` to not remove the `process` function. Does this have anything to do with the `"_rintf": True` parameter in `viewer.add_widget`?
            
Jupyter notebooks:
It first says that we will try to set up the notebooks and install the imjoy extension, but then it says that you don't need to do that. It's confusing. Does it mean that after the course we will need to do that but for the course and to make it quickly, we will use the Imjoy-Engine?

- "Let's first try to setup a local Jupyter notebook, and install the [imjoy-jupyter-extension](https://github.com/imjoy-team/imjoy-jupyter-extension). Instead of installing directly Jupyter notebooks, we recommend our a wrapper library called [ImJoy-Engine](https://github.com/imjoy-team/imjoy-engine). This allows us to fix some settings on the Jupyter server which can be useful for ImJoy, and you don't need to install imjoy-jupyter-extension separately."

I don't understand what you want to say here:
* " * Other software, web applications and websites can also support[imjoy-rpc](https://github.com/imjoy-team/imjoy-rpc) which will further make it accessible to the entire ImJoy ecosystem. This includes: ITK/VTK Viewer, Vizarr, Kaibu, ImageJ.JS and CellPose."